requirements.txt
requests>=2.28.0
pandas>=1.5.0
pytest>=7.0.0
python-dotenv>=0.20.0
pydantic>=1.9.0
aiohttp>=3.8.0
plotly>=5.0.0
dash>=2.0.0
PyYAML>=6.0
